refactor(code_app): Split large files into modular structure
- environment_manager.py: Extract executor and models
- repository_integration.py: Split into dataset creators and file handlers
- slurm_manager.py: Split into slurm/ module with separate components
- services/__init__.py: Update imports
