+++ apps/code_app/services/__init__.py
@@ -3,8 +3,8 @@
 # Author: ywatanabe
 # File: apps/code_app/services/__init__.py
 
-from .slurm_manager import SlurmManager
 from .project_service_manager import ProjectServiceManager
+from .slurm import SlurmManager
 
 __all__ = ["SlurmManager", "ProjectServiceManager"]
